+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for `/itchysats/rollover/3.0.0`. This fixes a bug where inverse payout curves where capped at double the value of the initial price.
-- Support for `/itchysats/collab-settlement/2.0.0`.
+- Support for `/itchysats/collab-settlement/2.0.0`. This fixes a bug where inverse payout curves where capped at double the value of the initial price.
 
 ### Changed